Update skhd keybinds
Use lalt instead of cmd for some bindings.
Using cmd for things like cmd + return to open a terminal meant it
wasn't possible to use it for things applications often use it for like
creating a newline instead of sending a message, or confirming text
boxes in applications like Sublime Merge.