Fix git using wrong email
Because it defaulted user.email to the account's email on the current
host, that means it used my work email for personal repos while on my
work laptop.

Instead, hardcode the default email to my personal one, and override it
for work projects manually. They are always in very specific directories
both at home and work anyway.