Improve yabai setup
yabai-helper is now an actual CLI instead of having to source it and
then run functions. This makes it possible to use it as a user, and also
looks nicer to use in configs since we can just call it like a normal
command.

Also changes the setup of spaces so it's more idempotent, and add a
possibility to set specific wallpapers for spaces.